requirements.txt
accelerate
teradataml
numpy
typing-extensions
pandas
matplotlib
scikit-learn
python-dotenv
sentence-transformers
huggingface-hub
chardet
swifter
bitsandbytes
streamlit
langid
joblib
xgboost
safetensors
transformers
huggingface_hub